+++ src/DoneLogger/Program.cs
@@ -15,7 +15,7 @@ static class Program
             var config = ConfigService.Load();
             var logService = new LogService(config);
             var trackingService = new TimeTrackingService(logService);
-            var summaryService = new SummaryService(logService, config);
+            var summaryService = new SummaryService(logService, config, trackingService);
             Application.Run(new MainForm(config, logService, trackingService, summaryService));
         }
         catch (Exception ex)